requirements.txt
opencv-python
pygetwindow
pyautogui
Pillow
pywin32
PyQt5